refactor(hub-common): move helpers to _internal
affects: @esri/hub-common
